fixed cli
