--- .env.example
@@ -1,8 +1,0 @@
-NODE_ENV=development
-DATABASE_URL=postgresql://postgres:postgres@localhost:5432/matt-init-example-project_db
-POSTGRES_DB=matt-init-example-project_db
-POSTGRES_USER=
-POSTGRES_PASSWORD=
-POSTGRES_PORT=5432
-BETTER_AUTH_SECRET=your_better_auth_secret # you can use openssl for this
-BETTER_AUTH_URL=your_better_auth_url
