+++ src/lib/db/schema/index.ts
@@ -1,1 +1,2 @@
 export * from "./auth";
+export * from "./game";
